+++ include/rain/networking/http/client.hpp
@@ -4,5 +4,6 @@
 #include "socket.hpp"
 
 namespace Rain::Networking::Http {
+	// TODO.
 	class Client : protected Http::Socket {};
 }