+++ test/socket-server.cpp
@@ -10,7 +10,8 @@ int main() {
 		std::cout << "onCloseSlave\n";
 	};
 	std::cout << "Starting server...\n";
-	server.serve(Rain::Networking::Host(NULL, 80), false);
+	server.serve(Rain::Networking::Host(NULL, 0), false);
+	std::cout << "Serving on port " << server.getService().getCStr() << "\n";
 	Rain::Time::sleepMs(10000);
 	std::cout << "Stopping server...\n";
 	server.close();